refactor(core): fold core internals into srp modules
