+++ crates/uselesskey-core-cache/Cargo.toml
@@ -17,6 +17,7 @@ authors.workspace = true
 [dependencies]
 dashmap = { workspace = true, optional = true }
 spin = { version = "0.10.0", default-features = false, features = ["mutex", "spin_mutex"] }
+uselesskey-core = { path = "../uselesskey-core", version = "0.6.0", default-features = false }
 uselesskey-core-id = { path = "../uselesskey-core-id", version = "0.6.0", default-features = false }
 
 [dev-dependencies]
@@ -27,7 +28,7 @@ serde.workspace = true
 
 [features]
 default = ["std"]
-std = ["dep:dashmap", "uselesskey-core-id/std"]
+std = ["dep:dashmap", "uselesskey-core/std", "uselesskey-core-id/std"]
 
 [package.metadata.docs.rs]
 features = ["std"]
