+++ crates/uselesskey-core-factory/Cargo.toml
@@ -16,6 +16,7 @@ authors.workspace = true
 
 [dependencies]
 rand10 = { workspace = true }
+uselesskey-core = { path = "../uselesskey-core", version = "0.6.0", default-features = false }
 uselesskey-core-cache = { path = "../uselesskey-core-cache", version = "0.6.0", default-features = false }
 uselesskey-core-id = { path = "../uselesskey-core-id", version = "0.6.0", default-features = false }
 
@@ -29,6 +30,7 @@ default = ["std"]
 std = [
     "rand10/std",
     "rand10/sys_rng",
+    "uselesskey-core/std",
     "uselesskey-core-cache/std",
     "uselesskey-core-id/std",
 ]
