+++ crates/uselesskey-core-hash/Cargo.toml
@@ -16,6 +16,7 @@ authors.workspace = true
 
 [dependencies]
 blake3.workspace = true
+uselesskey-core = { path = "../uselesskey-core", version = "0.6.0", default-features = false }
 
 [dev-dependencies]
 insta.workspace = true
@@ -24,7 +25,7 @@ serde.workspace = true
 
 [features]
 default = ["std"]
-std = ["blake3/std"]
+std = ["blake3/std", "uselesskey-core/std"]
 
 [package.metadata.docs.rs]
 features = ["std"]
