+++ crates/uselesskey-core-keypair/Cargo.toml
@@ -15,7 +15,7 @@ documentation = "https://docs.rs/uselesskey-core-keypair"
 authors.workspace = true
 
 [dependencies]
-uselesskey-core-keypair-material.workspace = true
+uselesskey-core = { path = "../uselesskey-core", version = "0.6.0" }
 
 [dev-dependencies]
 insta.workspace = true
