+++ crates/uselesskey-core-negative-der/Cargo.toml
@@ -15,6 +15,7 @@ documentation = "https://docs.rs/uselesskey-core-negative-der"
 authors.workspace = true
 
 [dependencies]
+uselesskey-core = { path = "../uselesskey-core", version = "0.6.0", default-features = false }
 uselesskey-core-hash = { workspace = true }
 
 [dev-dependencies]
@@ -25,7 +26,7 @@ serde.workspace = true
 
 [features]
 default = ["std"]
-std = ["uselesskey-core-hash/std"]
+std = ["uselesskey-core/std", "uselesskey-core-hash/std"]
 
 [package.metadata.docs.rs]
 features = ["std"]
