+++ crates/uselesskey-core-negative/Cargo.toml
@@ -15,6 +15,7 @@ documentation = "https://docs.rs/uselesskey-core-negative"
 authors.workspace = true
 
 [dependencies]
+uselesskey-core = { path = "../uselesskey-core", version = "0.6.0", default-features = false }
 uselesskey-core-negative-der = { workspace = true }
 uselesskey-core-negative-pem = { workspace = true }
 
@@ -26,7 +27,11 @@ serde.workspace = true
 
 [features]
 default = ["std"]
-std = ["uselesskey-core-negative-der/std", "uselesskey-core-negative-pem/std"]
+std = [
+    "uselesskey-core/std",
+    "uselesskey-core-negative-der/std",
+    "uselesskey-core-negative-pem/std",
+]
 
 [package.metadata.docs.rs]
 features = ["std"]
