+++ crates/uselesskey-core-seed/Cargo.toml
@@ -18,6 +18,7 @@ authors.workspace = true
 blake3 = { version = "1.8.4", default-features = false }
 rand_chacha10.workspace = true
 rand_core10.workspace = true
+uselesskey-core = { path = "../uselesskey-core", version = "0.6.0", default-features = false }
 
 [dev-dependencies]
 insta.workspace = true
@@ -27,7 +28,7 @@ serde.workspace = true
 
 [features]
 default = ["std"]
-std = ["blake3/std", "rand_chacha10/std"]
+std = ["blake3/std", "rand_chacha10/std", "uselesskey-core/std"]
 
 [package.metadata.docs.rs]
 features = ["std"]
