+++ crates/uselesskey-core-sink/Cargo.toml
@@ -16,6 +16,7 @@ authors.workspace = true
 
 [dependencies]
 tempfile.workspace = true
+uselesskey-core = { path = "../uselesskey-core", version = "0.6.0" }
 
 [dev-dependencies]
 insta.workspace = true
