+++ crates/uselesskey-core/Cargo.toml
@@ -17,28 +17,33 @@ authors.workspace = true
 [features]
 default = ["std"]
 std = [
-    "dep:uselesskey-core-sink",
+    "dep:base64",
+    "dep:dashmap",
+    "dep:tempfile",
+    "blake3/std",
+    "rand10/std",
+    "rand10/sys_rng",
+    "rand_chacha10/std",
     "thiserror/std",
-    "uselesskey-core-factory/std",
-    "uselesskey-core-cache/std",
-    "uselesskey-core-id/std",
-    "uselesskey-core-negative/std",
 ]
 
 [dependencies]
+base64 = { workspace = true, optional = true }
+blake3.workspace = true
+dashmap = { workspace = true, optional = true }
+rand10.workspace = true
+rand_chacha10.workspace = true
+rand_core10.workspace = true
+spin = { version = "0.10.0", default-features = false, features = ["mutex", "spin_mutex"] }
+tempfile = { workspace = true, optional = true }
 thiserror = { version = "2.0.18", default-features = false }
-uselesskey-core-factory = { path = "../uselesskey-core-factory", version = "0.6.0", default-features = false }
-uselesskey-core-sink = { path = "../uselesskey-core-sink", version = "0.6.0", optional = true }
-
-uselesskey-core-cache = { path = "../uselesskey-core-cache", version = "0.6.0", default-features = false }
-uselesskey-core-id = { path = "../uselesskey-core-id", version = "0.6.0", default-features = false }
-uselesskey-core-negative = { path = "../uselesskey-core-negative", version = "0.6.0", default-features = false }
 
 [dev-dependencies]
 insta.workspace = true
 proptest.workspace = true
 rstest.workspace = true
 serde = { workspace = true }
+uselesskey-test-support = { workspace = true }
 
 [package.metadata.docs.rs]
 features = ["std"]
