+++ crates/uselesskey-core/src/negative/mod.rs
@@ -1,1 +1,4 @@
-pub use uselesskey_core_negative::*;
+pub use crate::srp::negative::{
+    CorruptPem, corrupt_der_deterministic, corrupt_pem, corrupt_pem_deterministic, flip_byte,
+    truncate_der,
+};
