+++ crates/uselesskey-core/src/srp/keypair.rs
@@ -1,0 +1,3 @@
+//! Shared keypair compatibility module.
+
+pub use crate::srp::keypair_material::Pkcs8SpkiKeyMaterial;
