+++ crates/uselesskey-core/src/srp/mod.rs
@@ -1,0 +1,14 @@
+//! Internal single-responsibility modules for core fixture mechanics.
+
+pub mod cache;
+pub mod factory;
+pub mod hash;
+pub mod identity;
+#[cfg(feature = "std")]
+pub mod keypair;
+#[cfg(feature = "std")]
+pub mod keypair_material;
+pub mod negative;
+pub mod seed;
+#[cfg(feature = "std")]
+pub mod sink;
