+++ crates/uselesskey-ecdsa/Cargo.toml
@@ -16,7 +16,6 @@ authors.workspace = true
 
 [dependencies]
 uselesskey-core = { path = "../uselesskey-core", version = "0.6.0" }
-uselesskey-core-keypair-material.workspace = true
 
 # Elliptic curve support from RustCrypto's current RC line
 p256 = { version = "0.14.0-rc.8", features = ["ecdsa", "pkcs8", "pem"] }
