+++ crates/uselesskey-rsa/Cargo.toml
@@ -16,7 +16,6 @@ authors.workspace = true
 
 [dependencies]
 uselesskey-core = { path = "../uselesskey-core", version = "0.6.0" }
-uselesskey-core-keypair-material.workspace = true
 rsa.workspace = true
 rsa09 = { package = "rsa", version = "0.9.10", features = ["pem"], optional = true }
 rand_chacha = { version = "0.3.1", default-features = false, optional = true }
