+++ crates/uselesskey-rsa/src/keypair.rs
@@ -22,8 +22,8 @@ use rsa10::pkcs8::DecodePublicKey;
 use rsa10::pkcs8::{EncodePrivateKey, EncodePublicKey, LineEnding};
 use uselesskey_core::negative::CorruptPem;
 use uselesskey_core::sink::TempArtifact;
+use uselesskey_core::srp::keypair_material::Pkcs8SpkiKeyMaterial;
 use uselesskey_core::{Error, Factory};
-use uselesskey_core_keypair_material::Pkcs8SpkiKeyMaterial;
 
 use crate::RsaSpec;
